+++ test_v6.py
@@ -31,7 +31,7 @@ User: {instruction}
 Assistant:"""
 
 device = 'npu'
-model = Rwkv6ForCausalLM.from_pretrained("RWKV/rwkv-6-world-1b6", torch_dtype=torch.float16).to(device)
+model = Rwkv6ForCausalLM.from_pretrained("RWKV/rwkv-6-world-1b6", torch_dtype=torch.float32).to(device)
 tokenizer = Rwkv6Tokenizer.from_pretrained("RWKV/rwkv-6-world-1b6")
